Module wasmer_llvm_backend::backend

Structs

LLVMBackend

LLVMCache

Enums

Buffer

Statics

SIGNAL_HANDLER_INSTALLED

Functions

cxx_invoke_trampoline⚠

get_callbacks

get_func_symbol⚠

invoke_trampoline⚠
invoke_trampoline is a wrapper around cxx_invoke_trampoline, for fixing up the obsoleted trap_out in the C++ part.

llvm_backend_get_code_ptr⚠

llvm_backend_get_code_size⚠

llvm_backend_get_stack_map_ptr⚠

llvm_backend_get_stack_map_size⚠

module_delete⚠

module_load⚠

throw_any⚠
This should be the same as spliting up the fat pointer into two arguments, but this is cleaner, I think?

throw_breakpoint⚠

throw_trap⚠